New branch that uses DSA (Data Structure Analysis).
DSA is a pointer analysis originally from PoolAlloc project.

The  command takes an optional parameter  which
resolves all the indirect calls using the points-to information
provided by DSA. This is done first by inferring for each function
pointer the set of possible targets. Then, it performs a bitcode
transformation where each indirect call is translated to a switch
statement where each case corresponds to a target.

Note that the points-to information can be used for specializing in
other situations. For now, we only use it to resolve indirect calls
but its use can be extended as needed.